% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{cereal_details}
\alias{cereal_details}
\title{Find needed details for vctrs prototype}
\usage{
cereal_details(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A list
}
\description{
Find needed details for vctrs prototype
}
\examples{
cereal_details(factor(letters[1:5], labels = "letter"))
cereal_details(factor(LETTERS[3:1], ordered = TRUE))
cereal_details(as.POSIXct("2023-01-01", tz = "America/New_York"))
}
